refactor: 가맹점관리, 창고관리 VO, DTO 패키지 생성
