+++ src/common/util/InputUtil.java
@@ -1,3 +1,5 @@
+package common.util;
+
 import java.util.Optional;
 import java.util.Scanner;
