+++ src/common/util/MenuUtil.java
@@ -1,3 +1,5 @@
+package common.util;
+
 import java.util.Map;
 import java.util.Optional;
