+++ src/dto/franchise/FranchiseDTO.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.franchise;
 
 public class FranchiseDTO {
 }
