+++ src/dto/franchise/MinOrderDTO.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.franchise;
 
 public class MinOrderDTO {
 }
