+++ src/dto/franchise/ProductDTO.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.franchise;
 
 public class ProductDTO {
 }
