+++ src/dto/warehouse/WarehouseDTO.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.warehouse;
 
 public class WarehouseDTO {
 }
