+++ src/repository/InboundRepo.java
@@ -2,7 +2,6 @@ package repository;
 
 import dto.inbound.InboundDTO;
 import dto.inbound.ProductDTO;
-import vo.ProductVO;
 import vo.inbound.InboundVO;
 import vo.inbound.InboundDetailVO;
