+++ src/vo/franchise/ProductVO.java
@@ -1,9 +1,7 @@
-package vo;
+package vo.franchise;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-
-import java.util.Date;
 
 @Getter
 @AllArgsConstructor
